refactor db connection to be class based
